refactor(server): extract workspace detail assembly
